moved the determine result to the verifier-validator
